move Cluster abstract class to separate file
